+++ next/package.json
@@ -13,6 +13,7 @@
     "@sanity/client": "^6.11.1",
     "dayjs": "^1.11.10",
     "next": "14.1.0",
+    "next-mdx-remote": "^4.4.1",
     "next-sanity": "^7.0.10",
     "plop": "^4.0.1",
     "react": "^18",
