+++ next/src/components/_global/faq/faq.tsx
@@ -1,24 +1,12 @@
-'use client';
-
-import { useState } from 'react';
 import styles from './faq.module.scss';
 import { type Props } from './faq.types';
+import FaqItem from './faqItem';
 import Img from '@/components/ui/Img';
 import Markdown from '@/components/ui/Markdown';
 import Heading from '@/components/ui/heading/Heading';
 import CentralizedHeading from '@/components/ui/centralizedHeading/centralizedHeading';
 
 export default function Faq({ data: { centralizedHeading, description, faq, image, title, heading } }: Props) {
-  const [activeIndex, setActiveIndex] = useState(-1);
-
-  const handleButtonClick = (i: number) => {
-    if (i == activeIndex) {
-      setActiveIndex(-1);
-    } else {
-      setActiveIndex(i);
-    }
-  };
-
   return (
     <section className={styles.faqWrapper}>
       <div className={`${styles.faq} maxWidth`}>
@@ -38,25 +26,26 @@ export default function Faq({ data: { centralizedHeading, description, faq, imag
             data={image}
           />
           <div className={styles.items}>
-            {faq.map(({ title, description, image }, i) => (
-              <div
-                className={i == activeIndex ? `${styles.item} ${styles.active}` : styles.item}
-                key={i}
-              >
+            {faq.map(({ title, description, image }, i) => {
+              const arrow = <Arrow />;
+              const Image = (
                 <Img
                   className={styles.image}
                   data={image}
                 />
-                <Markdown.h3 className={styles.heading}>{title}</Markdown.h3>
-                <button
-                  className={styles.icon}
-                  onClick={() => handleButtonClick(i)}
-                >
-                  <Arrow />
-                </button>
-                <Markdown className={styles.description}>{description}</Markdown>
-              </div>
-            ))}
+              );
+
+              const Heading = <Markdown.h3 className={styles.heading}>{title}</Markdown.h3>;
+
+              const Description = <Markdown className={styles.description}>{description}</Markdown>;
+
+              return (
+                <FaqItem
+                  data={{ Image, Heading, Description, arrow, key: i }}
+                  key={i}
+                />
+              );
+            })}
           </div>
         </div>
         <CentralizedHeading data={centralizedHeading} />
