+++ next/src/components/ui/Markdown.tsx
@@ -1,129 +1,103 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-
+import { MDXRemote } from 'next-mdx-remote/rsc';
 import Link from 'next/link';
-import ReactMarkdown from 'react-markdown';
-const LinkRenderer = ({ href, children }) => {
-  const isExternal = href && (href.startsWith('https://') || href.startsWith('mailto:') || href.startsWith('tel:'));
-  return isExternal ? (
-    <a
-      href={href}
-      target='_blank'
-      rel='noopener noreferrer'
+import { domain } from '@/global/Seo';
+const LinkRenderer = ({
+  href,
+  children,
+}: React.AnchorHTMLAttributes<HTMLAnchorElement> & {
+  children?: React.ReactNode;
+}) => {
+  const isExternal =
+    href &&
+    ((href.startsWith('https://') && !href.startsWith(domain)) ||
+      href.startsWith('mailto:') ||
+      href.startsWith('tel:'));
+  const Element = isExternal ? 'a' : Link;
+  return (
+    <Element
+      href={href || ''}
       className='link'
+      {...(isExternal && {
+        target: '_blank',
+        rel: 'noopener',
+      })}
     >
       {children}
-    </a>
-  ) : (
-    <Link
-      href={href}
-      className='link'
-    >
-      {children}
-    </Link>
+    </Element>
   );
 };
-
-const ListRenderer = ({ children, ordered }) => (
+const ListRenderer = ({
+  children,
+  ordered,
+}: React.LiHTMLAttributes<HTMLLIElement> & {
+  children?: React.ReactNode;
+  ordered?: boolean;
+}) => (
   <li>
-    {!ordered && <ListBullet />}
+    {!ordered && <BulletList />}
     <span>{children}</span>
   </li>
 );
-
 type MarkdownProps = {
-  level?: any;
-  children?: any;
-  components?: any;
+  Tag?: keyof JSX.IntrinsicElements;
+  components?: Record<string, React.ReactNode>;
+  children: string;
   className?: string;
 };
-
-const Markdown = ({ level, children, components, className, ...props }: MarkdownProps) => {
-  const HeadingComponent = level;
-  const updatedComponents = level
-    ? {
-        ...components,
-        p: ({ children }) => <HeadingComponent {...props}>{children}</HeadingComponent>,
-      }
-    : components;
-
-  return (
-    <ReactMarkdown
+const Markdown = ({ Tag, components, children, className, ...props }: MarkdownProps) => {
+  const markdown = (
+    <MDXRemote
+      source={children}
       components={{
+        ...(Tag && {
+          p: ({ children }) => <Tag {...props}>{children}</Tag>,
+        }),
         a: LinkRenderer,
         li: ListRenderer,
         ol: ({ children }) => <ol className='orderedList'>{children}</ol>,
         ul: ({ children }) => <ul className='unorderedList'>{children}</ul>,
-        ...updatedComponents,
+        ...components,
       }}
-      className={className}
       {...props}
-    >
-      {children}
-    </ReactMarkdown>
+    />
   );
+  return className ? <div className={className}>{markdown}</div> : markdown;
 };
-
-Markdown.h1 = (props) => (
+Markdown.h1 = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h1'
+    Tag='h1'
     {...props}
   />
 );
-Markdown.h2 = (props) => (
+Markdown.h2 = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h2'
+    Tag='h2'
     {...props}
   />
 );
-Markdown.h3 = (props) => (
+Markdown.h3 = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h3'
+    Tag='h3'
     {...props}
   />
 );
-Markdown.h4 = (props) => (
+Markdown.h4 = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h4'
+    Tag='h4'
     {...props}
   />
 );
-Markdown.h5 = (props) => (
+Markdown.h5 = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h5'
+    Tag='h5'
     {...props}
   />
 );
-Markdown.h6 = (props) => (
+Markdown.span = (props: JSX.IntrinsicAttributes & MarkdownProps) => (
   <Markdown
-    level='h6'
+    Tag='span'
     {...props}
   />
 );
-
-const ListBullet = () => (
-  <svg
-    xmlns='http://www.w3.org/2000/svg'
-    width='24'
-    height='24'
-    fill='none'
-    viewBox='0 0 24 24'
-  >
-    <path
-      fill='currentColor'
-      d='M4 11.25a.75.75 0 000 1.5v-1.5zm0 1.5h16v-1.5H4v1.5z'
-      opacity='0.5'
-    ></path>
-    <path
-      stroke='currentColor'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-      strokeWidth='1.5'
-      d='M14 6l6 6-6 6'
-    ></path>
-  </svg>
-);
-
 export default Markdown;
+const BulletList = () => <svg></svg>;
